refactor(glens): extract shared types, constants, markdown helpers, formatElapsed
